[Fix/#9] WishlistItem model 분리
